+++ test/db/utilities/retail_offers/retail_suppliers_offers_archive_test.dart
@@ -3,6 +3,7 @@ library test.utilities.rate_board_test;
 import 'dart:convert';
 
 import 'package:date/date.dart';
+import 'package:elec_server/client/utilities/retail_offers/retail_supply_offer.dart';
 import 'package:elec_server/client/utilities/retail_suppliers_offers.dart';
 import 'package:elec_server/src/db/utilities/retail_suppliers_offers_archive.dart';
 import 'package:http/http.dart';
@@ -35,7 +36,7 @@ Future<void> tests(String rootUrl) async {
         'supplierName',
         'offerPostedOnDate',
       });
-      expect(x0['rate'], 0.1689);
+      expect(x0['rate'], 168.9);
     });
   });
 
@@ -43,13 +44,27 @@ Future<void> tests(String rootUrl) async {
   ///
   group('Retail suppliers offers client tests', () {
     var client = RetailSuppliersOffers(Client(), rootUrl: rootUrl);
+    var term = Term.parse('1Jan22-4Dec22', UTC);
+    late List<RetailSupplyOffer> offers;
+    setUp(() async {offers = await client.getOffersForRegionTerm('ISONE', term);});
     test('get ISONE offers', () async {
-      var term = Term.parse('1Jan22-4Dec22', UTC);
-      var offers = await client.getOffersForRegionTerm('ISONE', term);
       var x0 = offers.firstWhere((e) => e.offerId == 'ct-49486');
-      expect(x0.rate, 0.1689);
+      expect(x0.rate, 168.9);
       expect(x0.supplierName, 'NRG Business');
     });
+
+    test('get current offers as of a given date', () {
+      var o1 = RetailSuppliersOffers.getCurrentOffers(offers, Date.utc(2022, 11, 17));
+      expect(o1
+          .where((e) => e.supplierName == 'Constellation NewEnergy, Inc.')
+          .toList().length, 0);
+      // as of 2022-11-28 there are new offers
+      var o2 = RetailSuppliersOffers.getCurrentOffers(offers, Date.utc(2022, 11, 28));
+      expect(o2
+          .where((e) => e.supplierName == 'Constellation NewEnergy, Inc.')
+          .toList().length, 4);
+    });
+
   });
 }